Moved function to utils folder
